adicionar autenticação JWT com implementação de middleware para bloqueio de rotas

+++ controllers/post.js
@@ -2,7 +2,6 @@
 const User = require('../models/user');
 // Importing Post model
 const Post = require('../models/post');
-const { where } = require('sequelize');
 
 // Creating post controlelrs
 const getAllPosts = async (req, res) => {
@@ -30,9 +29,18 @@ const getAllPosts = async (req, res) => {
 // Get a posts by userid
 const getPost = async (req, res) => {
     const {userid} = req.params;
+    const {sortBy='id', order='ASC'} = req.query;
     try{
         const posts = await Post.findAll({
-            where: {userid:userid}
+            where: {userid:userid},
+            order: [[sortBy, order.toUpperCase()]],
+            include: [
+                {
+                    model: User,
+                    as: 'author',
+                    attributes: ['name']
+                }
+            ]
         });
         res.status(200).json(posts);
     } catch (err){
@@ -45,6 +53,13 @@ const getPost = async (req, res) => {
 // Create new Post by User id
 const postUserPost = async (req, res) => {
     const {description, img_data=null, userid} = req.body;
+    // Statement to block a different user to create a Post through another user
+    if (req.user.id != userid){
+        return res.status(403).json({
+            message: 'Action denied.'
+        });
+    }
+    // Creating new Post
     try{
         const newPost = await Post.create({description, img_data, userid});
         res.status(201).json({
@@ -62,6 +77,23 @@ const postUserPost = async (req, res) => {
 const deletePost = async (req, res) => {
     const {id} = req.params;
     try{
+        // Get the Post by Id provided
+        const post = await Post.findOne({
+            where: {id}
+        });
+        // Checks if the post by Id provided exist
+        if (!post){
+            return res.status(404).json({
+                message: 'Post does not exist.'
+            });
+        }
+        // Checks if the userid in the auth is equals to post.userid
+        if (req.user.id != post.userid){
+            return res.status(403).json({
+                message: 'Action denied.'
+            });
+        } 
+        // Delete the post from database
         await Post.destroy({
             where: {id:id}
         });
